+++ packages/mwp-app-render/src/util/browserPolyfill.js
